Rename config file to be aligned with templates
