+++ src/api/scopes/controllers/admin/create-scope.js
@@ -1,7 +1,7 @@
 import Joi from 'joi'
 import Boom from '@hapi/boom'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { createScope } from '~/src/api/scopes/helpers/create-scope.js'
 import { scopeNameExists } from '~/src/api/scopes/helpers/scope-name-exists.js'
