+++ src/api/scopes/controllers/admin/delete-scope.js
@@ -1,6 +1,6 @@
 import Boom from '@hapi/boom'
 import Joi from '~/src/helpers/extended-joi.js'
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 
 import { scopeExists } from '~/src/api/scopes/helpers/scope-exists.js'
 import { deleteScope } from '~/src/helpers/mongo/transactions/delete-scope.js'
