+++ src/api/scopes/controllers/admin/update-scope.js
@@ -1,7 +1,7 @@
 import Joi from '~/src/helpers/extended-joi.js'
 import Boom from '@hapi/boom'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { updateScope } from '~/src/api/scopes/helpers/update-scope.js'
 import { scopeExists } from '~/src/api/scopes/helpers/scope-exists.js'
