+++ src/api/teams/controllers/add-user-to-team.js
@@ -1,6 +1,6 @@
 import Boom from '@hapi/boom'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { getTeam } from '~/src/api/teams/helpers/get-team.js'
 import { getUser } from '~/src/api/users/helpers/get-user.js'
 import { teamHasUser } from '~/src/api/teams/helpers/team-has-user.js'
