+++ src/api/teams/controllers/create-team.js
@@ -1,6 +1,6 @@
 import Boom from '@hapi/boom'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { createTeamValidationSchema } from '~/src/api/teams/helpers/create-team-validation-schema.js'
 import { MongoErrors } from '~/src/helpers/mongodb-errors.js'
 import { teamNameExists } from '~/src/api/teams/helpers/team-name-exists.js'
