+++ src/api/teams/controllers/delete-team.js
@@ -1,6 +1,6 @@
 import Joi from 'joi'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { deleteTeam } from '~/src/helpers/mongo/transactions/delete-transactions.js'
 import { removeUserFromAadGroup } from '~/src/api/teams/helpers/remove-user-from-aad-group.js'
