+++ src/api/teams/controllers/update-team.js
@@ -1,6 +1,6 @@
 import Boom from '@hapi/boom'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { updateTeamValidationSchema } from '~/src/api/teams/helpers/update-team-validation-schema.js'
 import { getTeam } from '~/src/api/teams/helpers/get-team.js'
 import { getTeamsCount } from '~/src/api/teams/helpers/get-teams.js'
