+++ src/api/teams/helpers/github/github-shared-repo-access.js
@@ -1,4 +1,4 @@
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { createLogger } from '~/src/helpers/logging/logger.js'
 
 const logger = createLogger()
