+++ src/api/teams/helpers/github/github-team-exists.js
@@ -1,4 +1,4 @@
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 
 async function gitHubTeamExists(octokit, github) {
   const org = config.get('gitHubOrg')
