+++ src/api/teams/helpers/github/search-github-teams.js
@@ -1,4 +1,4 @@
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 
 async function searchGitHubTeams(octokit, query) {
   const org = config.get('gitHubOrg')
