+++ src/api/teams/helpers/group-name-from-team-name.js
@@ -1,4 +1,4 @@
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 
 function groupNameFromTeamName(teamName) {
   const groupPrefix = config.get('azureGroupPrefix')
