+++ src/api/users/controllers/create-user.js
@@ -1,6 +1,6 @@
 import Boom from '@hapi/boom'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { createUserValidationSchema } from '~/src/api/users/helpers/create-user-validation-schema.js'
 import { MongoErrors } from '~/src/helpers/mongodb-errors.js'
 import { aadUserIdExists } from '~/src/api/users/helpers/aad-user-id-exists.js'
