+++ src/api/users/helpers/search-github-users.js
@@ -1,4 +1,4 @@
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 
 async function searchGitHubUsers(octokit, query) {
   const org = config.get('gitHubOrg')
