+++ src/helpers/azure-oidc.js
@@ -1,6 +1,6 @@
 import jwt from '@hapi/jwt'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { proxyFetch } from '~/src/helpers/proxy.js'
 
 const azureOidc = {
