+++ src/helpers/docs/swagger-options.js
@@ -1,4 +1,4 @@
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 
 const service = config.get('service')
