+++ src/helpers/ms-graph.js
@@ -2,7 +2,7 @@ import { ClientSecretCredential } from '@azure/identity'
 import { Client } from '@microsoft/microsoft-graph-client'
 import { TokenCredentialAuthenticationProvider } from '@microsoft/microsoft-graph-client/authProviders/azureTokenCredentials/index.js'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { provideProxy } from '~/src/helpers/proxy.js'
 
 const msGraphPlugin = {
