+++ src/helpers/proxy.test.js
@@ -1,7 +1,7 @@
 import fetchMock from 'jest-fetch-mock'
 import { ProxyAgent } from 'undici'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { provideProxy, proxyFetch } from '~/src/helpers/proxy.js'
 
 const mockLoggerDebug = jest.fn()
